+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Examples to README.rst
+### Fixed
+- Wrong super() argument in fs.proxy.writer
+- fs.proxy._utils.__all__ containing an undefined name cause by
+  renaming of `unique_everseen` to unique
 
 ## [v0.1.1] - 2017-06-29
 ### Fixed